Changed ids to be of the long data type instead of int

+++ src/main/java/io/brightskies/loyalty/order/repo/OrderRepo.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import io.brightskies.loyalty.order.entity.Order;
 
-public interface OrderRepo extends JpaRepository<Order, Integer> {
+public interface OrderRepo extends JpaRepository<Order, Long> {
 
 }
